backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pillow>=10.0.0
PyMuPDF>=1.23.0
python-multipart>=0.0.6
